arm64: dts: uniphier: change release-addr of spin-table to normal memory
Mark Rutland pointed out that it is not safe to use MMIO registers for
spin-table.

Mark Rutland wrote:
Generally, it is not safe to use MMIO registers to back spin-table. The
kernel maps the spin table location with cacheable attributes, so there
may be speculative accesses to any registes in the same (64K) page, and
a writeback may be larger than the 8-byte register width (which the
device might not accept, triggering an SError).

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>