i2c: uniphier-f: fix misdetection of incomplete stop condition
Currently, the status register FI2C_SR is checked immediately after
a stop condition is issued.  It takes typically 5-10 usec until the
corresponding bits in the register are set, so the error check for
"stop condition was not completed" is always false positive.

Add wait code to relax the status register check.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>