fbdev: simplefb: add support 'memory-region' property on devicetree node
This patch enables 'memory-region' property that refers to memory description
on reserved-memory node in the case of devicetree use.
When there isn't a 'reg' property that specifies the address and the size of
the framebuffer, the address and the size writen in the memory description on
reserved-memory node are used for the framebuffer.

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>