[tentative] spi: add the generic compatible string for spidev
Without the generic compatible string (ex."spidev"), the spidev occurs
WARN_ON while booting.
This issue was discussed by the following thread:
   https://patchwork.kernel.org/patch/6113191/

The regular way is to add the device-specific compatible string for
the connected device.

Since this is a tentative patch, we need to change the way after that.

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>