[backport 499fef09a323] reset: uniphier: Add glue layer reset control
Add a reset line to enable glue layer implemented in UniPhier SoCs.

This reuses only the reset operations in reset-simple, because
the reset-simple doesn't handle any SoC-dependent clocks and resets.
This reset line is included in the glue layer, and it's necessary
to enable clocks and deassert resets of the layer before using this
reset line.

This commit is a combination of the following upstreamed commits.
  commit 499fef09a323 ("reset: uniphier: add USB3 core reset control")
  commit 3eb8f765f5ae ("reset: uniphier-usb3: Rename to reset-uniphier-glue")

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>