[backport cd8843f5] i2c: uniphier-f: fill TX-FIFO only in IRQ handler for repeated START
commit cd8843f541cc0ef057e27f1afba46374bbb84219 upstream.

- For a repeated START condition, this controller starts data transfer
immediately after the slave address is written to the TX-FIFO.

- Once the TX-FIFO empty interrupt is asserted, the controller makes
a pause even if additional data are written to the TX-FIFO.

Given those circumstances, the data after a repeated START may not be
transferred if the interrupt is asserted while the TX-FIFO is being
filled up. A more reliable way is to append TX data only in the
interrupt handler.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>